requirements.txt
async-fastapi-jwt-auth==0.6.4
SQLAlchemy[asyncio]==2.0.29
alembic== 1.13.1
redis==5.0.3
typer==0.12.2
Werkzeug==3.0.2
pydantic==2.6.4
uvicorn==0.29.0
gunicorn==21.2.0
ruff==0.3.5
mypy==1.9.0